\name{x}
\alias{x}
\docType{data}
\title{
The toy data example for input data matrix x in the example
}
\description{
This is the p by n toy data we used to illustrate the package.
}
\usage{data("x")}

\details{
The toy data is a p by n matrix, where n is the sample size
}
\references{
Set-based differential covariance testing for genomics
}
\author{
Yi-Hui Zhou
}
\examples{
data(x)
dim(x)
}
\keyword{datasets}
